+++ CHANGELOG.md
@@ -69,6 +69,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Renamed /warewulf/wwinit to /warewulf/prescripts. #1156
 - Display auto-detected kernel version during iPXE and GRUB. #1742
 - Reduced default verbosity of `wwctl overlay build`.
+- Reorganized and updated userdocs. #1116, #1324
 
 ### Fixed